pyproject.toml
[project]
name = "athena"
version = "0.1.0"
description = ""
authors = [
    {name = "piyushk6626",email = "kulkarnipiyush462@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "async-timeout (==5.0.1)",
    "attrs (==25.1.0)",
    "beautifulsoup4 (==4.12.3)",
    "bs4 (==0.0.2)",
    "certifi (==2024.12.14)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.1)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "distro (==1.9.0)",
    "exceptiongroup (==1.2.2)",
    "fastapi (==0.115.8)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "jiter (==0.8.2)",
    "openai (==1.61.0)",
    "outcome (==1.3.0.post0)",
    "packaging (==24.2)",
    "pinecone (==6.0.0)",
    "pinecone-plugin-interface (==0.0.7)",
    "pydantic (==2.10.6)",
    "pydantic-core (==2.27.2)",
    "pysocks (==1.7.1)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "redis (==5.2.1)",
    "requests (==2.32.3)",
    "selenium (>=4.30.0,<5.0.0)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sortedcontainers (==2.4.0)",
    "soupsieve (==2.6)",
    "spotipy (==2.25.0)",
    "starlette (==0.45.3)",
    "tqdm (==4.67.1)",
    "trio (==0.28.0)",
    "trio-websocket (==0.11.1)",
    "typing-extensions (==4.12.2)",
    "undetected-chromedriver (==3.5.5)",
    "urllib3 (==2.3.0)",
    "uvicorn (==0.34.0)",
    "webdriver-manager (==4.0.2)",
    "websocket-client (==1.8.0)",
    "websockets (==14.2)",
    "wsproto (==1.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
